+++ test/unit/msw-mock/handlers/deletable-element-handler.ts
@@ -6,4 +6,7 @@ export const deletableElementHandlers: RestHandler[] = [
   rest.delete(`http://localhost/${ElementUuid.DeletableElement}`, (_req, res, ctx) => {
     return res(ctx.status(204));
   }),
+  rest.delete(`http://localhost/${ElementUuid.DataNode}`, (_req, res, ctx) => {
+    return res(ctx.status(204));
+  }),
 ];