+++ CHANGELOG.md
@@ -21,3 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add [MSW](https://mswjs.io/) for better unit tests, wip
 - Improve error and log messages
 - Remove yarn.lock, as this is a library
+- Add functions EmberNexus.getElement(), .patchElement(), .putElement() and .deleteElement(), add tests for these
+  functions
+- Update GitHub actions